requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
pydantic
requests
pdfminer.six
python-docx
groq
